# stockone/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stock.one API",
        "description": "API de gestion de stocks pour papeteries - Benin",
        "contact": {
            "email": "admin@stockone.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://stockone.test:81/api/v1",
            "description": "Serveur local Laragon"
        }
    ],
    "paths": {
...
        "/exports/sales": {
            "get": {
                "tags": [
                    "Export"
                ],
                "summary": "Exporter les ventes en CSV",
                "description": "Export ventes CSV",
                "operationId": "0f6b3eec197f9b99e954523670dfdeab",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "format",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "csv",
                            "enum": [
                                "csv",
                                "excel"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier telecharge"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/exports/sales/details": {
            "get": {
                "tags": [
                    "Export"
                ],
                "summary": "Exporter le detail des ventes en CSV",
                "description": "Export lignes de ventes détaillées CSV",
                "operationId": "02fe7e10236dd1324f5145ac08d92836",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier telecharge"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/exports/stock": {
            "get": {
                "tags": [
                    "Export"
                ],
                "summary": "Exporter l'etat du stock en CSV",
                "description": "Export stock CSV",
                "operationId": "a7b5d191b76cb9d36322a9dc59e4200f",
                "responses": {
                    "200": {
                        "description": "Fichier telecharge"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/exports/credits": {
            "get": {
                "tags": [
                    "Export"
                ],
                "summary": "Exporter les credits en CSV",
                "description": "Export crédits CSV",
                "operationId": "89897236d1c692b9a17a014456f358bc",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "all",
                            "enum": [
                                "all",
                                "pending",
                                "partial",
                                "overdue",
                                "doubtful"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier telecharge"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/exports/stock/movements": {
            "get": {
                "tags": [
                    "Export"
                ],
                "summary": "Exporter les mouvements de stock en CSV",
                "description": "Export mouvements de stock CSV",
                "operationId": "cca231dcb10977fbd249a68d3e361daa",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier telecharge"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}